backend/requirements.txt
fastapi==0.103.2
uvicorn==0.22.0
python-dotenv==1.0.0
groq==0.26.0
pydantic==2.4.2
python-multipart==0.0.6